+++ src/entropy.jl
@@ -1,11 +1,16 @@
-export EntropyDefinition, DiffEntropyEst, DifferentialEntropyEstimator
+export EntropyDefinition
+export MLEntropy, DiscEntropyEst, DiscreteEntropyEstimator
+export DiffEntropyEst, DifferentialEntropyEstimator
 export entropy, entropy_maximum, entropy_normalized
 
 """
     EntropyDefinition
 
 `EntropyDefinition` is the supertype of all types that encapsulate definitions
-of (generalized) entropies. Currently implemented entropy definitions are:
+of (generalized) entropies. These also serve as estimators of discrete entropies,
+see description below.
+
+Currently implemented entropy definitions are:
 
 - [`Renyi`](@ref).
 - [`Tsallis`](@ref).
@@ -14,11 +19,7 @@ of (generalized) entropies. Currently implemented entropy definitions are:
 - [`Curado`](@ref).
 - [`StretchedExponential`](@ref).
 
-These entropy types are given as inputs to [`entropy`](@ref) and
-[`entropy_normalized`](@ref). Notice that in all documentation strings formulas are
-provided for the discrete version of the entropy, for simplicity.
-
-See [`entropy`](@ref) for usage.
+These types can be given as inputs to [`entropy`](@ref) or [`entropy_normalized`](@ref).
 
 ## Description
 
@@ -26,64 +27,74 @@ Mathematically speaking, generalized entropies are just nonnegative functions of
 probability distributions that verify certain (entropy-type-dependent) axioms.
 Amigó et al.'s[^Amigó2018] summary paper gives a nice overview.
 
+However, for a software implementation computing entropies _in practice_,
+definitions is not really what matters; **estimators matter**.
+Because in the practical sense, one needs to estimate a definition from finite data,
+and different ways of estimating a quantity come with their own pros and cons.
+
+That is why the type [`DiscreteEntropyEstimator`](@ref) exists,
+which is what is actually given to [`entropy`](@ref).
+Some ways to estimate a discrete entropy only apply to a specific entropy definition.
+For estimators that can be applied to various entropy definitions,
+this is specified by providing an instance of `EntropyDefinition` to the estimator.
+
 [^Amigó2018]:
     Amigó, J. M., Balogh, S. G., & Hernández, S. (2018). A brief review of
     generalized entropies. [Entropy, 20(11), 813.](https://www.mdpi.com/1099-4300/20/11/813)
 """
 abstract type EntropyDefinition end
 
-"""
-    DifferentialEntropyEstimator
-    DiffEntropyEst # alias
-
-The supertype of all differential entropy estimators.
-These estimators compute an entropy value in various ways that do not involve
-explicitly estimating a probability distribution.
-
-Currently implemented estimators are:
-
-- [`KozachenkoLeonenko`](@ref)
-- [`Kraskov`](@ref)
-- [`Zhu`](@ref)
-- [`ZhuSingh`](@ref)
-- [`Goria`](@ref)
-- [`Gao`](@ref)
-- [`Lord`](@ref)
-- [`Vasicek`](@ref)
-- [`Ebrahimi`](@ref)
-- [`Correa`](@ref)
-- [`AlizadehArghami`](@ref)
-
-See [`entropy`](@ref) for usage.
-"""
-abstract type DifferentialEntropyEstimator end
-const DiffEntropyEst = DifferentialEntropyEstimator
-
 ###########################################################################################
 # Discrete entropy
 ###########################################################################################
-# Notice that StatsBase.jl also exports `entropy`.
 """
-    entropy([e::EntropyDefinition,] probs::Probabilities)
-    entropy([e::EntropyDefinition,] est::ProbabilitiesEstimator, x)
+    DiscreteEntropyEstimator
+    DiscEntropyEst # alias
 
-Compute the **discrete entropy** `h::Real ∈ [0, ∞)` defined by `e`, in one of two ways:
+Supertype of all discrete entropy estimators.
+
+Currently only the [`MLEntropy`](@ref) estimator is provided,
+which does not need to be used, as using an [`EntropyDefinition`](@ref) directly in
+[`entropy`](@ref) is possible. But in the future, more advanced estimators will
+be added ([#237](https://github.com/JuliaDynamics/ComplexityMeasures.jl/issues/237)).
+"""
+abstract type DiscreteEntropyEstimator end
+const DiscEntropyEst = DiscreteEntropyEstimator
+
+# Dummy estimator that doesn't actually change anything from the definitions
+"""
+    MLEntropy(e::EntropyDefinition) <: DiscreteEntropyEstimator
+
+Standing for "maximum likelihood entropy", and also called empirical/naive/plug-in,
+this estimator calculates the entropy exactly as defined in the given
+[`EntropyDefinition`](@ref) directly from a probability mass function.
+"""
+struct MLEntropy{E<:EntropyDefinition} <: DiscreteEntropyEstimator
+    definition::E
+end
+
+# Notice that StatsBase.jl also exports `entropy`!
+"""
+    entropy([e::DiscreteEntropyEstimator,] probs::Probabilities)
+    entropy([e::DiscreteEntropyEstimator,] est::ProbabilitiesEstimator, x)
+
+Compute the **discrete entropy** `h::Real ∈ [0, ∞)`,
+using the estimator `e`, in one of two ways:
 
 1. Directly from existing [`Probabilities`](@ref) `probs`.
-2. From input data `x`, by first estimating a probability distribution using the provided
-   [`ProbabilitiesEstimator`](@ref), then computing entropy from that distribution.
-   In fact, the second method is just a 2-lines-of-code wrapper that calls
-   [`probabilities`](@ref) and gives the result to the first method.
+2. From input data `x`, by first estimating a probability mass function using the provided
+   [`ProbabilitiesEstimator`](@ref), and then computing the entropy from that mass fuction
+   using the provided [`DiscreteEntropyEstimator`](@ref).
 
-The entropy definition (first argument) is optional. Explicitly provide `e` if you need to
-specify a logarithm base for the entropy. When `est` is a probability estimator,
-`Shannon()` is used by default.
+Instead of providing a [`DiscreteEntropyEstimator`](@ref), an [`EntropyDefinition`](@ref)
+can be given directly, in which case [`MLEntropy`](@ref) is used as the estimator.
+If `e` is not provided, [`Shannon`](@ref)`()` is used by default.
 
 ## Maximum entropy and normalized entropy
 
-All discrete entropies `e` have a well defined maximum value for a given probability estimator.
-To obtain this value one only needs to call the [`entropy_maximum`](@ref) function with the
-chosen entropy type and probability estimator. Or, one can use [`entropy_normalized`](@ref)
+All discrete entropies have a well defined maximum value for a given probability estimator.
+To obtain this value one only needs to call the [`entropy_maximum`](@ref).
+Or, one can use [`entropy_normalized`](@ref)
 to obtain the normalized form of the entropy (divided by the maximum).
 
 ## Examples
@@ -103,30 +114,48 @@ function entropy(e::EntropyDefinition, est::ProbabilitiesEstimator, x)
     return entropy(e, ps)
 end
 
-# dispatch for `entropy(e, ps)` in the entropy definitions files
+# dispatch for `entropy(e::EntropyDefinition, ps::Probabilities)`
+# is in the individual entropy definitions files
 
 # Convenience
 entropy(est::ProbabilitiesEstimator, x) = entropy(Shannon(), est, x)
 entropy(probs::Probabilities) = entropy(Shannon(), probs)
+entropy(e::MLEntropy, args...) = entropy(e.definition, args...)
 
 ###########################################################################################
 # Differential entropy
 ###########################################################################################
+"""
+    DifferentialEntropyEstimator
+    DiffEntropyEst # alias
+
+The supertype of all differential entropy estimators.
+These estimators compute an entropy value in various ways that do not involve
+explicitly estimating a probability distribution.
+
+See the [table of differential entropy estimators](@ref table_diff_ent_est)
+in the docs for all differential entropy estimators.
+
+See [`entropy`](@ref) for usage.
+"""
+abstract type DifferentialEntropyEstimator end
+const DiffEntropyEst = DifferentialEntropyEstimator
+
 # Dispatch for these functions is implemented in individual estimator files in
 # `entropies/estimators/`.
 """
-    entropy([e::EntropyDefinition,] est::DifferentialEntropyEstimator, x)
+    entropy(est::DifferentialEntropyEstimator, x) → h
 
-Approximate the **differential entropy** `h::Real` according to the definition `e` using
-the provided [`DifferentialEntropyEstimator`](@ref) and input data `x`.
+Approximate the **differential entropy** `h::Real` using the provided
+[`DifferentialEntropyEstimator`](@ref) and input data `x`.
 This method doesn't involve explicitly computing (discretized) probabilities first.
 
-The entropy definition argument is optional.
-The default entropy type is inferred from the estimator (e.g. [`Kraskov`](@ref)
-estimates the base-2 Shannon differential entropy).
-The estimators are not compatible with all versions of [`EntropyDefinition`](@ref).
-See [Table of differential entropy estimators](@ref)
-in the docs for a table view of the estimators and the compatibilities.
+The overwhelming majority of entropy estimators estimate the Shannon entropy.
+If some estimator can estimate different _definitions_ of entropy (e.g., [`Tsallis`](@ref)),
+this is provided as an argument to the estimator itself.
+
+See the [table of differential entropy estimators](@ref table_diff_ent_est)
+in the docs for all differential entropy estimators.
 
 ## Examples
 
@@ -134,40 +163,37 @@ A standard normal distribution has a base-e differential entropy of `0.5*log(2π
 nats.
 
 ```julia
-def = Shannon(; base = ℯ) # Base `ℯ` for nats.
-est = Kraskov(k = 5)
-h = entropy(def, est, randn(1_000_000))
+est = Kraskov(k = 5, base = ℯ) # Base `ℯ` for nats.
+h = entropy(est, randn(1_000_000))
 abs(h - 0.5*log(2π) - 0.5) # ≈ 0.001
 ```
-
 """
-function entropy(e::EntropyDefinition, est::DiffEntropyEst, x)
-    t = string(nameof(typeof(e)))
-    throw(ArgumentError("$t entropy not implemented for $(typeof(est)) estimator"))
-end
+function entropy(::DiffEntropyEst, ::Any) end
 
-entropy(est::DiffEntropyEst, ::Probabilities) =
-    error("EntropyDefinition estimators like $(nameof(typeof(est))) are not called with probabilities.")
-entropy(e::EntropyDefinition, est::DiffEntropyEst, ::Probabilities) =
-    error("EntropyDefinition estimators like $(nameof(typeof(est))) are not called with probabilities.")
-entropy(e::EntropyDefinition, est::DiffEntropyEst, x::AbstractVector) =
-    entropy(e, est, Dataset(x))
-# Always default to Shannon with base-2 logs. Individual estimators may override this.
-entropy(est::DiffEntropyEst, x) = entropy(Shannon(), est, x)
-
+entropy(est::DiffEntropyEst, ::Probabilities) = error("""
+    EntropyDefinition estimators like $(nameof(typeof(est)))
+    are not called with probabilities.
+""")
 
 ###########################################################################################
 # Normalize API
 ###########################################################################################
 """
-    entropy_maximum(e::EntropyDefinition, est::ProbabilitiesEstimator)
+    entropy_maximum(e::EntropyDefinition, est::ProbabilitiesEstimator, x)
 
-Return the maximum value of a discrete entropy the given probabilities estimator.
+Return the maximum value of a discrete entropy with the given probabilities estimator
+and input data `x`. Like in [`outcome_space`](@ref), for some estimators
+the concrete outcome space is known without knowledge of input `x`,
+in which case the function dispatches to `entropy_maximum(e, est)`.
 
     entropy_maximum(e::EntropyDefinition, L::Int)
 
-Same as above, but computed thdirectly from the number of total outcomes `L`.
+Same as above, but computed directly from the number of total outcomes `L`.
 """
+function entropy_maximum(e::EntropyDefinition, est::ProbabilitiesEstimator, x)
+    L = total_outcomes(est, x)
+    return entropy_maximum(e, L)
+end
 function entropy_maximum(e::EntropyDefinition, est::ProbabilitiesEstimator)
     L = total_outcomes(est)
     return entropy_maximum(e, L)
@@ -175,24 +201,29 @@ end
 function entropy_maximum(e::EntropyDefinition, ::Int)
     error("not implemented for entropy type $(nameof(typeof(e))).")
 end
+entropy_maximum(e::MLEntropy, args...) = entropy_maximum(e.definition, args...)
 
 """
-    entropy_normalized([e::EntropyDefinition,] est::ProbabilitiesEstimator, x) → h̃
+    entropy_normalized([e::DiscreteEntropyEstimator,] est::ProbabilitiesEstimator, x) → h̃
 
 Return `h̃ ∈ [0, 1]`, the normalized discrete entropy of `x`, i.e. the value of [`entropy`](@ref)
 divided by the maximum value for `e`, according to the given probabilities estimator.
-If `e` is not given, it defaults to `Shannon()`.
+
+Instead of a discrete entropy estimator, an [`EntropyDefinition`](@ref)
+can be given as first argument. If `e` is not given, it defaults to `Shannon()`.
 
 Notice that there is no method
-`entropy_normalized(e::EntropyDefinition, probs::Probabilities)`, because there is no way to know
+`entropy_normalized(e::DiscreteEntropyEstimator, probs::Probabilities)`,
+because there is no way to know
 the amount of _possible_ events (i.e., the [`total_outcomes`](@ref)) from `probs`.
 """
 function entropy_normalized(e::EntropyDefinition, est::ProbabilitiesEstimator, x)
-    return entropy(e, est, x) / entropy_maximum(e, est)
+    return entropy(e, est, x) / entropy_maximum(e, est, x)
 end
 function entropy_normalized(est::ProbabilitiesEstimator, x::Array_or_Dataset)
     return entropy_normalized(Shannon(), est, x)
 end
+entropy_normalized(e::MLEntropy, est, x) = entropy_normalized(e.definition, est, x)
 
 ###########################################################################################
 # Utils