+++ test/entropies/estimators/kozachenkoleonenko.jl
@@ -9,17 +9,7 @@ N = round(0.5*log(2π) + 0.5, digits = 2)
 N_base3 = round((0.5*log(2π) + 0.5) / log(3, ℯ), digits = 2) # custom base
 
 npts = 1000000
-ea = entropy(Shannon(), KozachenkoLeonenko(), rand(npts))
-ea_n = entropy(Shannon(; base = ℯ), KozachenkoLeonenko(), randn(npts))
-ea_n3 = entropy(Shannon(; base = 3), KozachenkoLeonenko(), randn(npts))
-
+ea = entropy(KozachenkoLeonenko(), rand(npts))
+ea_n3 = entropy(KozachenkoLeonenko(base = 3), randn(npts))
 @test U - max(0.01, U*0.03) ≤ ea ≤ U + max(0.01, U*0.03)
-@test N * 0.98 ≤ ea_n ≤ N * 1.02
 @test N_base3 * 0.98 ≤ ea_n3 ≤ N_base3 * 1.02
-
-x = rand(1000)
-@test_throws ArgumentError entropy(Renyi(q = 2), KozachenkoLeonenko(), x)
-
-# Default is Shannon base-2 differential entropy
-est = KozachenkoLeonenko()
-@test entropy(est, x) == entropy(Shannon(), est, x)