+++ test/entropies/estimators/vasicek.jl
@@ -1,3 +1,4 @@
+using ComplexityMeasures, Test
 # -------------------------------------------------------------------------------------
 # Check if the estimator converge to true values for some distributions with
 # analytically derivable entropy.
@@ -7,20 +8,10 @@ U = 0.00
 # EntropyDefinition with natural log of 𝒩(0, 1) is 0.5*ln(2π) + 0.5.
 N = round(0.5*log(2π) + 0.5, digits = 2)
 N_base3 = round((0.5*log(2π) + 0.5) / log(3, ℯ), digits = 2) # custom base
-N_base2 = round((0.5*log(2π) + 0.5) / log(2, ℯ), digits = 2) # custom base
 
 npts = 1000000
-ea = entropy(Shannon(), Vasicek(m = 100), rand(npts))
-ea_n = entropy(Shannon(; base = ℯ), Vasicek(m = 100), randn(npts))
-ea_n3 = entropy(Shannon(; base = 3), Vasicek(m = 100), randn(npts))
+ea = entropy(Vasicek(m = 100), rand(npts))
+ea_n3 = entropy(Vasicek(m = 100, base = 3), randn(npts))
 
 @test U - max(0.02, U*0.03) ≤ ea ≤ U + max(0.01, U*0.03)
-@test N * 0.96 ≤ ea_n ≤ N * 1.02
 @test N_base3 * 0.96 ≤ ea_n3 ≤ N_base3 * 1.02
-
-x = rand(1000)
-@test_throws ArgumentError entropy(Renyi(q = 2), Vasicek(), x)
-
-# Default is Shannon base-2 differential entropy
-est = Vasicek()
-@test entropy(est, x) == entropy(Shannon(), est, x)