+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Changelog is kept with respect to version 0.11 of Entropies.jl. From version v2.0 onwards, this package has been renamed to ComplexityMeasures.jl.
 
+## 2.3
+- Like differential entropies, discrete entropies now also have their own estimator type.
+- The approach of giving both an entropy definition, and an entropy estimator to `entropy` has been dropped. Now the entropy estimators know what definitions they are applied for. This change is a deprecation, i.e., backwards compatible.
+- Added `MLEntropy` discrete entropy estimator.
+
 ## 2.2
 
 - Corrected documentation for `SymbolicPermutation`, `SymbolicAmplitudeAwarePermutation`,